Extract Planner from MongoDbPlanner into separate file
